Use ClassImporter mixin
