+++ package.json
@@ -75,6 +75,7 @@
     "ldapjs": "git+https://github.com/schul-cloud/node-ldapjs.git",
     "lodash": "^4.17.13",
     "migrate-mongoose": "^4.0.0",
+    "mixwith": "^0.1.1",
     "moment": "^2.19.2",
     "mongoose": "^4.7.6",
     "mongoose-autopopulate": "^0.9.1",
